requirements.txt
# Cotton Weed Detection Project Requirements
# Install with: pip install -r requirements.txt

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO
ultralytics>=8.0.0

# Data
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0

# Images
opencv-python>=4.7.0
pillow>=9.5.0

# Cleaning / quality
cleanlab>=2.2.0
scikit-learn>=1.2.0

# Plotting
matplotlib>=3.6.0
